+++ README.md
@@ -1,5 +1,7 @@
 # music-automaton
-A node app that automatically generates 16 measures long melodies according to certain harmonic rules.
+A JavaScript library to automatically generate small melodies based on a set of predefined rules.
+
+The output is a [MusicXML](https://www.musicxml.com/) string. MusicXML files can be imported into most composition softwares like [MuseScore](https://musescore.org/en) and there are other JavaScript libraries that can then be used to for example display MusicXML on a website.
 
 ## Running the application
