+++ tsconfig.json
@@ -3,18 +3,18 @@
         "target": "es6",
         "module": "commonjs",
         "outDir": "dist",
-        "sourceMap": true
+        "sourceMap": true,
+        "declaration": true
     },
     "files": [
         "./node_modules/@types/jest/index.d.ts",
         "./node_modules/@types/node/index.d.ts"
     ],
     "include": [
-        "app.ts",
         "src/**/*.ts"
     ],
     "exclude": [
         "node_modules"
-    ]
+    ],
 }
